requirements.txt
django_phonenumber_field==2.1.0
Django==2.1.3
django_localflavor==2.1
djangorestframework==3.9.0
django-crispy-forms==1.7.2
phonenumbers==8.10.0